refactor oui module
